+++ spec/rubocop/cop/style/symbol_proc_spec.rb
@@ -84,6 +84,12 @@ describe RuboCop::Cop::Style::SymbolProc, :config do
     expect(cop.offenses).to be_empty
   end
 
+  it 'accepts block with splat params' do
+    inspect_source(cop, 'something { |*x| x.first }')
+
+    expect(cop.offenses).to be_empty
+  end
+
   context 'when the method has arguments' do
     let(:source) { 'method(one, 2) { |x| x.test }' }